requirements.txt
azure-common==1.1.28
azure-core==1.24.1
azure-identity==1.10.0
azure-keyvault-secrets==4.4.0
Babel==2.10.2
blinker==1.4
cachelib==0.8.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-AdminLTE3==1.0.8
Flask-BabelEx==0.9.4
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.4.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
oauth2==1.9.0.post1
oauthlib==3.2.0
passlib==1.7.4
portalocker==2.4.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.37
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
